fitur user deskripsi produk catalog done
fitur user deskripsi produk catalog done

+++ database/migrations/2025_11_05_074928_add_is_active_to_categories_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('categories', function (Blueprint $table) {
-            $table->boolean('is_active')->default(true);
+            if (!Schema::hasColumn('categories', 'is_active')) {
+                $table->boolean('is_active')->default(true);
+            }
         });
     }
 
@@ -22,7 +24,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('categories', function (Blueprint $table) {
-            $table->dropColumn('is_active');
+            if (Schema::hasColumn('categories', 'is_active')) {
+                $table->dropColumn('is_active');
+            }
         });
     }
 };
